+++ packages/nimble-components/src/anchor-tab/tests/anchor-tab.react.tsx
@@ -1,4 +1,0 @@
-import { AnchorTab } from '..';
-import { wrap } from '../../utilities/tests/react-wrapper';
-
-export const NimbleAnchorTab = wrap(AnchorTab);