+++ packages/nimble-components/src/menu-button/tests/menu-button.react.tsx
@@ -1,4 +1,0 @@
-import { MenuButton } from '..';
-import { wrap } from '../../utilities/tests/react-wrapper';
-
-export const NimbleMenuButton = wrap(MenuButton);