+++ packages/storybook/src/nimble/select/select-opened-in-div.stories.ts
@@ -1,0 +1,44 @@
+import type { StoryFn, Meta } from '@storybook/html';
+import { html, ViewTemplate } from '@microsoft/fast-element';
+import { listOptionTag } from '@ni/nimble-components/dist/esm/list-option';
+import { selectTag } from '@ni/nimble-components/dist/esm/select';
+import { createStory } from '../../utilities/storybook';
+import { sharedMatrixParameters } from '../../utilities/matrix';
+
+const metadata: Meta = {
+    title: 'Tests/Select',
+    parameters: {
+        ...sharedMatrixParameters()
+    }
+};
+
+export default metadata;
+
+const positionStates = [
+    ['below', ''],
+    ['above', 'margin-top: 120px;']
+] as const;
+type PositionState = (typeof positionStates)[number];
+
+// prettier-ignore
+const component = ([
+    position,
+    positionStyle
+]: PositionState): ViewTemplate => html`
+    <div style=${() => (position === 'below' ? 'height: 150px' : null)}>
+        <div style="overflow: auto; border: 2px solid red; ${() => positionStyle}">
+            <${selectTag} open position="${() => position}">
+                <${listOptionTag} value="1">Option 1</${listOptionTag}>
+                <${listOptionTag} value="2">Option 2</${listOptionTag}>
+                <${listOptionTag} value="3">Option 3</${listOptionTag}>
+            </${selectTag}>
+        </div>
+    </div>
+`;
+
+export const selectBelowNotConfinedByDiv: StoryFn = createStory(
+    component(positionStates[0])
+);
+export const selectAboveNotConfinedByDiv: StoryFn = createStory(
+    component(positionStates[1])
+);