+++ packages/storybook/src/nimble/table-column/icon/table-column-icon.react.tsx
@@ -1,0 +1,4 @@
+import { TableColumnIcon } from '@ni/nimble-components/dist/esm/table-column/icon';
+import { wrap } from '../../../utilities/react-wrapper';
+
+export const NimbleTableColumnIcon = wrap(TableColumnIcon);